+++ test/GPU/runtests.jl
@@ -9,3 +9,7 @@ include("spectral_transform.jl")
 
 # kernels
 include("kernels_GPU.jl")
+
+# test if the models run on GPU 
+include("barotropic.jl")
+